Fixed migrations and db:seed
So we can rebuild database from a fresh checkout
Also, finish merging conflict in ProcessAccounts and scope accounts
migration

+++ database/migrations/2015_06_17_201326_add_scope_to_datasets_table.php
@@ -27,9 +27,11 @@ class AddScopeToDatasetsTable extends Migration {
 	 */
 	public function down()
 	{
-        Schema::table('accounts', function(Blueprint $table)
+        Schema::table('datasets', function(Blueprint $table)
                       {
-                      $table->dropColumn('scope');
+                          if (Schema::hasColumn('datasets', 'scope')){
+                                         $table->dropColumn('scope');
+                          }
                       });
 	}
